Add files via upload
FEW_calc is the Initial setup of classes and global factors from week 6 and FEW_gameplay is the code that will have the gameplay steps. At this stage I am not sure best approach for the game, especially some of the less straightforward aspects like how the water trackers have to move three spaces before they actually drop a value and the table of how surface water and current fish populations affect the affect to the fish population in the following round.